packages/shared/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lex-intel-shared"
version = "0.1.0"
description = "Shared models, schemas, and utilities for LexIntel"
requires-python = ">=3.9"
dependencies = [
    "sqlalchemy[asyncio]>=2.0",
    "pydantic>=2.0",
    "redis>=5.0",
]

[project.optional-dependencies]
extraction = [
    "PyPDF2>=3.0.1",
    "python-docx>=0.8.11",
]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
